Added 'data' to the TD elements.
This is a significant step. I modified the code to make use of
jQuery.data(), which adds arbitrary data to an element. The arbitary
data I add: the 'letter' itself.

http://api.jquery.com/jQuery.data/

I also added an 'id' attribute to each TD element, using this:

http://stackoverflow.com/questions/2176986/jquery-add-id-instead-of-class

I had this code earlier in the 'for y' loop, but I don't think it's
all that useful in there. In the end, the ids probably aren't helpful
now that I have the jQuery data working.

Finally, while struggling to get this working, I learned that the "..."
syntax in my CoffeeScript for-loops means 'exclude the last element.'