Enabled additional turns! Updated highlighting.
Now the game lets you pick another word after you have made a
selection!

The game highlighting is now different. Selections are in yellow,
and successful words are in green. This change was done in the
stylesheet and the TD element classes ('highlighted' instead of
'green').

I also decided to abandon the animation for now.