Added start of drag-select feature.
I want users to be able to select words by dragging the mouse. The
code changes of this commit contain the start of this implementation.
I followed these URLs:

http://stackoverflow.com/questions/4127118/can-you-detect-dragging-in-jquery
http://jsfiddle.net/W7tvD/

I have to refactor the code logic so that it is done in the drag in
addition to the mouse click actions.